+++ src/Rule/Attribute/Course/CourseTaxonomy.php
@@ -4,21 +4,25 @@
 
 namespace srag\Plugins\SrLifeCycleManager\Rule\Attribute\Course;
 
-use srag\Plugins\SrLifeCycleManager\Rule\Attribute\TaxonomyHelper;
+use srag\Plugins\SrLifeCycleManager\Rule\Attribute\TaxonomyAttribute;
 use ilDBInterface;
 use ilObjCourse;
+use ilObject;
 
 /**
  * @author Thibeau Fuhrer <thibeau@sr.solutions>
  */
-class CourseTaxonomy extends CourseAttribute
+class CourseTaxonomy extends TaxonomyAttribute
 {
-    use TaxonomyHelper;
+    /**
+     * @var ilDBInterface
+     */
+    protected $database;
 
     /**
-     * @var string[]
+     * @var ilObjCourse
      */
-    protected $taxonomies;
+    private $object;
 
     /**
      * @param ilDBInterface $database
@@ -26,39 +30,23 @@ class CourseTaxonomy extends CourseAttribute
      */
     public function __construct(ilDBInterface $database, ilObjCourse $course)
     {
-        parent::__construct($course);
-
-        $this->taxonomies = $this->getTaxonomies(
-            $database,
-            $course->getId()
-        );
+        $this->database = $database;
+        $this->object = $course;
     }
 
     /**
      * @inheritDoc
      */
-    public function getComparableValueTypes() : array
+    protected function getDatabase(): ilDBInterface
     {
-        return [
-            self::COMPARABLE_VALUE_TYPE_ARRAY,
-            self::COMPARABLE_VALUE_TYPE_STRING,
-        ];
+        return $this->database;
     }
 
     /**
      * @inheritDoc
      */
-    public function getComparableValue(string $type)
+    protected function getObject(): ilObject
     {
-        switch ($type) {
-            case self::COMPARABLE_VALUE_TYPE_ARRAY:
-                return $this->taxonomies;
-
-            case self::COMPARABLE_VALUE_TYPE_STRING:
-                return implode(',', $this->taxonomies);
-
-            default:
-                return null;
-        }
+        return $this->object;
     }
 }