+++ src/Rule/Attribute/Course/CourseTutors.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace srag\Plugins\SrLifeCycleManager\Rule\Attribute\Course;
+
+use ilParticipants;
+
+/**
+ * @author Thibeau Fuhrer <thibeau@sr.solutions>
+ */
+class CourseTutors extends CourseParticipantAttribute
+{
+    /**
+     * @inheritDoc
+     */
+    protected function getConsideredMembers(ilParticipants $participants): array
+    {
+        return $participants->getTutors();
+    }
+}