+++ src/Rule/Attribute/Group/GroupParticipants.php
@@ -1,0 +1,23 @@
+<?php declare(strict_types=1);
+
+/* Copyright (c) 2022 Thibeau Fuhrer <thibeau@sr.solutions> Extended GPL, see docs/LICENSE */
+
+namespace srag\Plugins\SrLifeCycleManager\Rule\Attribute\Group;
+
+use ilGroupParticipants;
+use ilObjGroup;
+use ilParticipants;
+
+/**
+ * @author Thibeau Fuhrer <thibeau@sr.solutions>
+ */
+class GroupParticipants extends GroupParticipantAttribute
+{
+    /**
+     * @inheritDoc
+     */
+    protected function getConsideredMembers(ilParticipants $participants): array
+    {
+        return $participants->getParticipants();
+    }
+}